+++ languages/ts/ts-schema-typebox-adapter/tsconfig.lib.json
@@ -1,7 +1,6 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
     "declaration": true,
     "types": ["node"]
   },
