+++ languages/ts/ts-schema/package.json
@@ -22,7 +22,7 @@
     "dist"
   ],
   "dependencies": {
-    "jtd-utils": "workspace:*",
+    "@arrirpc/type-defs": "workspace:*",
     "scule": "^1.3.0",
     "uncrypto": "^0.1.3"
   },
