+++ languages/ts/ts-schema/src/compiler/serialize.ts
@@ -15,7 +15,7 @@ import {
     type SchemaFormRef,
     type SchemaFormType,
     type SchemaFormValues,
-} from "jtd-utils";
+} from "@arrirpc/type-defs";
 import { camelCase } from "scule";
 
 import { type TemplateInput } from "./common";
