+++ package.json
@@ -74,7 +74,7 @@
     "start-server-and-test": "^2.0.9",
     "typescript": "5.7.2",
     "typescript-eslint": "^8.18.2",
-    "unbuild": "^3.0.1",
+    "unbuild": "^3.1.0",
     "vite": "6.0.6",
     "vitest": "2.1.8",
     "zod": "^3.24.1"
