+++ tooling/codegen-utils/package.json
@@ -22,10 +22,8 @@
     "dist"
   ],
   "dependencies": {
-    "jtd-utils": "workspace:*",
+    "@arrirpc/type-defs": "workspace:*",
     "scule": "^1.3.0"
   },
-  "devDependencies": {
-    "@arrirpc/schema": "workspace:*"
-  }
+  "devDependencies": {}
 }
