+++ tooling/codegen-utils/src/helpers.test.ts
@@ -1,8 +1,4 @@
-import { a } from "@arrirpc/schema";
-
 import {
-    type AppDefinition,
-    createAppDefinition,
     removeDisallowedChars,
     type RpcDefinition,
     setNestedObjectProperty,
@@ -149,120 +145,4 @@ describe("String utils", () => {
             expect(stringStartsWithNumber(input)).toBe(false);
         }
     });
-});
-
-test("create app definition", () => {
-    const SettingsParams = a.object(
-        {
-            userId: a.string(),
-        },
-        {
-            id: "SettingsParams",
-        },
-    );
-    const Settings = a.object(
-        {
-            colorScheme: a.enumerator(["SYSTEM", "LIGHT", "DARK"]),
-        },
-        {
-            id: "Settings",
-        },
-    );
-
-    const result = createAppDefinition({
-        procedures: {
-            sayHello: {
-                transport: "http",
-                method: "post",
-                path: "/say-hello",
-            },
-            createConnection: {
-                transport: "ws",
-                path: "/ws",
-                params: a.object({
-                    message: a.string(),
-                }),
-                response: a.object({
-                    message: a.string(),
-                }),
-            },
-            "utils.getSettings": {
-                transport: "http",
-                method: "get",
-                path: "/utils/get-settings",
-                params: SettingsParams,
-                response: Settings,
-            },
-        },
-    });
-    const expectedResult: AppDefinition = {
-        schemaVersion: "0.0.7",
-        procedures: {
-            sayHello: {
-                transport: "http",
-                method: "post",
-                path: "/say-hello",
-                params: undefined,
-                response: undefined,
-            },
-            createConnection: {
-                transport: "ws",
-                path: "/ws",
-                params: "CreateConnectionParams",
-                response: "CreateConnectionResponse",
-            },
-            "utils.getSettings": {
-                transport: "http",
-                method: "get",
-                path: "/utils/get-settings",
-                params: "SettingsParams",
-                response: "Settings",
-            },
-        },
-        definitions: {
-            CreateConnectionParams: {
-                properties: {
-                    message: {
-                        type: "string",
-                        metadata: {},
-                    },
-                },
-                metadata: {},
-            },
-            CreateConnectionResponse: {
-                properties: {
-                    message: {
-                        type: "string",
-                        metadata: {},
-                    },
-                },
-                metadata: {},
-            },
-            SettingsParams: {
-                properties: {
-                    userId: {
-                        type: "string",
-                        metadata: {},
-                    },
-                },
-                metadata: {
-                    id: "SettingsParams",
-                },
-            },
-            Settings: {
-                properties: {
-                    colorScheme: {
-                        enum: ["SYSTEM", "LIGHT", "DARK"],
-                        metadata: {},
-                    },
-                },
-                metadata: {
-                    id: "Settings",
-                },
-            },
-        },
-    };
-    expect(JSON.parse(JSON.stringify(result))).toStrictEqual(
-        JSON.parse(JSON.stringify(expectedResult)),
-    );
 });
