+++ tooling/codegen-utils/tsconfig.spec.json
@@ -14,6 +14,7 @@
     "src/**/*.spec.js",
     "src/**/*.test.jsx",
     "src/**/*.spec.jsx",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts",
+    "../type-defs/src/appDef.test.ts"
   ]
 }
