+++ tooling/json-schema-to-atd/package.json
@@ -22,6 +22,6 @@
     "dist"
   ],
   "dependencies": {
-    "jtd-utils": "workspace:*"
+    "@arrirpc/type-defs": "workspace:*"
   }
 }
