--- tooling/jtd-utils/README.md
@@ -1,11 +1,0 @@
-# jtd-utils
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build jtd-utils` to build the library.
-
-## Running unit tests
-
-Run `nx test jtd-utils` to execute the unit tests via [Vitest](https://vitest.dev).
