+++ tooling/type-defs/README.md
@@ -1,0 +1,3 @@
+# type-defs
+
+Internal library for Arri RPC containing some shared common type definitions and helpers.
