+++ tooling/type-defs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "jtd-utils",
+  "name": "@arrirpc/type-defs",
   "type": "module",
   "license": "MIT",
   "author": {
@@ -12,7 +12,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/modiimedia/arri.git",
-    "directory": "tooling/jtd-utils"
+    "directory": "tooling/type-defs"
   },
   "main": "./dist/index.cjs",
   "module": "./dist/index.mjs",
@@ -21,5 +21,7 @@
     "dist"
   ],
   "version": "0.69.2",
-  "dependencies": {}
+  "dependencies": {
+    "scule": "^1.3.0"
+  }
 }
