+++ tooling/type-defs/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./appDef";
+export * from "./typeDef";
