+++ tooling/type-defs/vite.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-    cacheDir: "../../node_modules/.vite/@arrirpc/typebox-adapter",
+    cacheDir: "../../node_modules/.vite/@arrirpc/eslint-plugin",
 
     // plugins: [
     //     viteTsConfigPaths({
     //         root: "../../",
-    //     }) as any,
+    //     }),
     // ],
 
     // Uncomment this if you are using workers.
@@ -19,16 +19,14 @@ export default defineConfig({
     // },
 
     test: {
+        globals: true,
+        reporters: ["default"],
         pool: "threads",
+        passWithNoTests: true,
         poolOptions: {
             threads: {
                 singleThread: true,
             },
-        },
-        globals: true,
-        reporters: ["default"],
-        cache: {
-            dir: "../../node_modules/.vitest",
         },
         environment: "node",
         include: ["src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"],
